Created a new abstract base class (LodBoxFbxIOBase) for the Importer and Exporter classes so they have a common constructor.
Converted keys of FBX_VERSION dictionary to be ints as that is allowed!

Converted the FBX_FORMAT dictionary to an Enum as that seems clearer for me as a fixed choice for formats.

Renamed the export_fbx function name to be less misleading

Upgraded more type hints.

Started on merging scenes from different LOD FBX files